backend/requirements.txt
fastapi
uvicorn
supabase
google-generativeai
python-dotenv
httpx<0.28
pytest
